# openapi/dist/aep-v1-m1.bundle.yaml
openapi: 3.1.0
info:
  title: Agent Enterprise Protocol M1
  version: 1.0.0-m1
  description: M1 contract for model discovery, authorization, and gateway connection.
  license:
    name: Apache 2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local M1 deployment
  - url: https://enterprise.example.com
    description: Protected deployment
security:
  - bearerAuth: []
...
paths:
...
  /aep/v1/auth/methods:
    get:
      tags:
        - Authentication
      operationId: getAuthenticationMethods
      security: []
      parameters:
        - name: deploymentHint
          in: query
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Available authentication methods
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationMethods'
        '404':
          $ref: '#/components/responses/Problem'
  /aep/v1/auth/password/login:
    post:
      tags:
        - Authentication
      operationId: loginWithPassword
      description: Authenticates an administrator-provisioned ZhiYuan platform account. HTTP and HTTPS are valid in development and production. HTTPS is strongly recommended outside a trusted private network because plain HTTP exposes credentials and bearer tokens in transit.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordLoginRequest'
      responses:
        '200':
          $ref: '#/components/responses/TokenResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /aep/v1/auth/password/change:
    post:
      tags:
        - Authentication
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '200':
          description: Rotated AEP session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
  /aep/v1/auth/federated/start:
    post:
      tags:
        - Authentication
      operationId: startFederatedLogin
      description: Starts a system-browser login through a customer OIDC provider or another server-side identity adapter.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FederatedLoginStartRequest'
      responses:
        '200':
          description: Federated login transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FederatedLoginStartResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /aep/v1/auth/exchange:
    post:
      tags:
        - Authentication
      operationId: exchangeFederatedAuthorizationCode
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FederatedExchangeRequest'
      responses:
        '200':
          $ref: '#/components/responses/TokenResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
  /aep/v1/auth/refresh:
    post:
      tags:
        - Authentication
      operationId: refreshAuthenticationSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          $ref: '#/components/responses/TokenResponse'
        '401':
          $ref: '#/components/responses/Problem'
  /aep/v1/auth/logout:
    post:
      tags:
        - Authentication
      operationId: logoutAuthenticationSession
      security:
        - aepBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Session revoked
        '401':
          $ref: '#/components/responses/Problem'
  /aep/v1/user/activation:
    post:
      tags:
        - Authentication
      operationId: activateDeploymentLicense
      description: Exchanges a signed deployment license for a short-lived Control Service entitlement token. The service verifies the complete signed envelope with its configured vendor public keys and never receives license signing private keys. Clients MUST verify the envelope locally before calling this operation.
      security:
        - aepBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LicenseActivationRequest'
      responses:
        '200':
          description: Deployment entitlement token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitlementTokenResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /aep/v1/user/me:
    get:
      tags:
        - User Runtime
      operationId: getCurrentIdentity
      responses:
        '200':
          description: Current identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentIdentity'
        '401':
          $ref: '#/components/responses/Problem-2'
  /aep/v1/user/heartbeat:
    post:
      tags:
        - User Control Events
      operationId: reportUserHeartbeat
      parameters:
        - $ref: '#/components/parameters/AepProtocolVersion'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatRequest'
      responses:
        '200':
          description: Liveness accepted and discovery metadata returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeartbeatResponse'
        '400':
          $ref: '#/components/responses/Problem-2'
        '401':
          $ref: '#/components/responses/Problem-2'
  /aep/v1/user/control-events:
    get:
      tags:
        - User Control Events
      operationId: listUserControlEvents
      description: Reading does not consume an event. Unacknowledged deliveries may be returned again.
      parameters:
        - $ref: '#/components/parameters/AepProtocolVersion'
        - $ref: '#/components/parameters/RequestId'
        - name: afterCursor
          in: query
          schema:
            type:
              - string
              - 'null'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Applicable unacknowledged control events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlEventPage'
        '401':
          $ref: '#/components/responses/Problem-2'
...
  /aep/v1/user/skills/manifest:
    get:
      tags:
        - User Runtime
      operationId: getSkillManifest
      parameters:
        - name: If-None-Match
          in: header
          schema:
            type: string
      responses:
        '200':
          description: Complete desired Skill manifest
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillManifest'
        '304':
          description: Manifest unchanged
        '401':
          $ref: '#/components/responses/Problem-2'
  /aep/v1/user/skills/{skillId}/versions/{version}/package:
    get:
      tags:
        - User Runtime
      operationId: downloadSkillPackage
      parameters:
        - $ref: '#/components/parameters/SkillId'
        - $ref: '#/components/parameters/Version'
      responses:
        '200':
          description: Skill ZIP package
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '403':
          $ref: '#/components/responses/Problem-2'
        '404':
          $ref: '#/components/responses/Problem-2'
  /aep/v1/user/skills/sync-results:
    post:
      tags:
        - User Runtime
      operationId: reportSkillSyncResult
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillSyncResult'
      responses:
        '202':
          description: Result accepted
        '400':
          $ref: '#/components/responses/Problem-2'
  /aep/v1/user/events/batch:
    post:
      tags:
        - User Runtime
      operationId: uploadEventBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventBatchRequest'
      responses:
        '200':
          description: Per-event result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventBatchResponse'
        '400':
          $ref: '#/components/responses/Problem-2'
        '413':
          $ref: '#/components/responses/Problem-2'
...
  /aep/v1/admin/users:
    get:
      tags:
        - Admin Users
      operationId: listPlatformUsers
      security:
        - aepBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit-2'
      responses:
        '200':
          description: Platform user page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformUserPage'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    post:
      tags:
        - Admin Users
      operationId: createPlatformUser
      security:
        - aepBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlatformUserRequest'
      responses:
        '201':
          description: Platform user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformUser'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /aep/v1/admin/users/import:
    post:
      tags:
        - Admin Users
      operationId: importPlatformUsers
      description: Imports at most 1000 administrator-provisioned accounts.
      security:
        - aepBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportPlatformUsersRequest'
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPlatformUsersResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /aep/v1/admin/users/{userId}:
    patch:
      tags:
        - Admin Users
      operationId: updatePlatformUser
      security:
        - aepBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdatePlatformUserRequest'
      responses:
        '200':
          description: Platform user updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformUser'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /aep/v1/admin/users/{userId}/reset-password:
    post:
      tags:
        - Admin Users
      operationId: resetPlatformUserPassword
      security:
        - aepBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Temporary password set and refresh sessions revoked
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /aep/v1/admin/skills:
    get:
      tags:
        - Admin Skills
      operationId: listSkills
      responses:
        '200':
          description: Skill list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSkillList'
    post:
      tags:
        - Admin Skills
      operationId: createSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillWrite'
      responses:
        '201':
          $ref: '#/components/responses/AdminSkill'
        '409':
          $ref: '#/components/responses/Problem-2'
  /aep/v1/admin/skills/{skillId}:
    parameters:
      - $ref: '#/components/parameters/SkillId'
    get:
      tags:
        - Admin Skills
      operationId: getSkill
      responses:
        '200':
          $ref: '#/components/responses/AdminSkill'
        '404':
          $ref: '#/components/responses/Problem-2'
    patch:
      tags:
        - Admin Skills
      operationId: updateSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillPatch'
      responses:
        '200':
          $ref: '#/components/responses/AdminSkill'
    delete:
      tags:
        - Admin Skills
      operationId: deleteSkill
      responses:
        '204':
          description: Skill withdrawn
...
  /aep/v1/admin/skill-assignments/{assignmentId}:
    delete:
      tags:
        - Admin Skills
      operationId: deleteSkillAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '204':
          description: Assignment removed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthenticationMethods:
      type: object
      description: Authentication methods enabled for the deployment. The built-in password method has the stable ID zhiyuan-password. Federated methods are optional and MUST NOT be returned unless a production identity adapter is configured.
      required:
        - deployment
        - methods
      properties:
        deployment:
          $ref: '#/components/schemas/Deployment'
        deploymentId:
          type: string
          minLength: 1
        preferredMethodId:
          type:
            - string
            - 'null'
          description: The ID of the preferred enabled method. Defaults to zhiyuan-password for the password-only production profile.
        methods:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AuthenticationMethod'
...
  responses:
    Problem:
      description: AEP Problem Details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...